requirements.txt
asgiref==3.2.3
boto3==1.12.15
botocore==1.15.15
django-annoying==0.10.6
django-appconf==1.0.3
django-imagekit==4.0.2
django-storages==1.9.1
djangorestframework==3.11.0
django==3.0.4
docutils==0.15.2
jmespath==0.9.5
pilkit==2.0
pillow==7.0.0
psycopg2-binary==2.8.4
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
